# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/moims/{id}/messages": {
      "post": {
        "operationId": "ChatController_send",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "메시지 전송 + 저장된 메시지 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "빈/과대 content — 400"
          },
          "401": {
            "description": "유효한 Supabase JWT 부재 — 401"
          },
          "403": {
            "description": "대상 모임의 멤버가 아님(또는 없는 모임) — 403"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      },
      "get": {
        "operationId": "ChatController_history",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "keyset 내림차순(최신순) 히스토리 + nextCursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 cursor — 400"
          },
          "401": {
            "description": "유효한 Supabase JWT 부재 — 401"
          },
          "403": {
            "description": "대상 모임의 멤버가 아님(또는 없는 모임) — 403"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    }
  },
  "info": {
    "title": "moyura API",
    "description": "moyura 모노레포 백엔드 API (인증 단계)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "메시지 본문(비어 있을 수 없음, 최대 2000자)",
            "example": "안녕하세요"
          }
        },
        "required": [
          "content"
        ]
      },
      "ChatMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "메시지 id(BigInt PK의 문자열 표현 — 단조 증가, keyset 커서 기준)",
            "example": "42"
          },
          "moimId": {
            "type": "string",
            "description": "메시지가 속한 모임 id",
            "example": "15ebe4ba-7f12-4e2c-bfa4-a0a9eb5022b8"
          },
          "senderId": {
            "type": "string",
            "description": "발신자 sub(= profile.id). 표시 이름은 멤버 목록에서 클라이언트가 해석한다.",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "content": {
            "type": "string",
            "description": "메시지 본문",
            "example": "안녕하세요"
          },
          "createdAt": {
            "type": "string",
            "description": "메시지 생성 시각(ISO-8601)",
            "example": "2026-06-14T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "moimId",
          "senderId",
          "content",
          "createdAt"
        ]
      },
      "ChatHistoryResponseDto": {
        "type": "object",
        "properties": {
          "messages": {
            "description": "커서 이전 최신순(내림차순) 메시지 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "다음 페이지 커서(가장 오래된 반환분의 id). 더 없으면 null.",
            "example": "37",
            "nullable": true
          }
        },
        "required": [
          "messages",
          "nextCursor"
        ]
      }
    }
  }
}